Don't always wait for operations to complete
It seems reasonable to me that the caller might want to implement their
own waiting/polling code - or more importantly have request code which
doesn't enforce that behaviour.
Its trivial to check for an incomplete status and call the waiting code
and run poll_api() so I'm separating request and wait.