Document running tests; add new tests
Expanding test coverage is helping me refresh my memory of how this
library works and what is involved.

Its also dragging up long lost memories of pytest so this is going to
be an interesting experience.

Some tests are broken but I haven't managed to get MagicMock to return
what I expect - problem for later debugging.